requirements.txt
fastapi
strawberry-graphql
sqlalchemy
pandas
streamlit
uvicorn
requests